Extract name generation into LlmCredential::NameGenerator
